fix: Only backup & restore ld.so.preload if present
/etc/ld.so.preload is not always present in the Raspberry Pi
foundations images. For example it is absent in the Bullseye arm64
image and therefore shall only backed up and restored if it it present
in the image. Otherwise the image build will fail due to `set -e`.

Same fix used in Commit 9b0eb93.

+++ install_debs_into_image.sh
@@ -136,7 +136,7 @@ fi
 # from ld.so:
 #   ERROR: ld.so: object '/usr/lib/arm-linux-gnueabihf/libarmmem-${PLATFORM}.so'
 #   from /etc/ld.so.preload cannot be preloaded (cannot open shared object file): ignored.
-mv "$IMAGEDIR/etc/ld.so.preload" "$IMAGEDIR/etc/ld.so.preload.bak"
+[[ -f "$IMAGEDIR/etc/ld.so.preload" ]] && mv "$IMAGEDIR/etc/ld.so.preload" "$IMAGEDIR/etc/ld.so.preload.bak"
 
 # customize settings
 echo `basename "$1"` > "$IMAGEDIR/etc/revpi/image-release"
@@ -158,7 +158,7 @@ find "$IMAGEDIR/var/log" -type f -delete
 find "$IMAGEDIR/etc/ssh" -name "ssh_host_*_key*" -delete
 
 # restore ld.so.preload
-mv "$IMAGEDIR/etc/ld.so.preload.bak" "$IMAGEDIR/etc/ld.so.preload"
+[[ -f "$IMAGEDIR/etc/ld.so.preload.bak" ]] && mv "$IMAGEDIR/etc/ld.so.preload.bak" "$IMAGEDIR/etc/ld.so.preload"
 
 cleanup_umount
